Move school input component to separate file
