+++ src/js/modules/core/components/ApplyPage.tsx
@@ -11,6 +11,7 @@ import { User } from "firebase";
 
 import { schools } from "./schools";
 import { db } from "../../../firebase";
+import SchoolInput from "./SchoolInput";
 
 
 interface Props {
@@ -34,71 +35,6 @@ interface ApplyPageState {
   isLoading: boolean;
 }
 
-interface InputState<T> {
-  value: T;
-}
-
-
-const getSuggestions = value => {
-  const inputValue = value.trim().toLowerCase();
-  const inputLength = inputValue.length;
-
-  return inputLength === 0 ? [] : languages.filter(lang =>
-    lang.name.toLowerCase().slice(0, inputLength) === inputValue
-  );
-};
-
-const isUpper = (str) => str === str.toUpperCase()
-
-class SchoolInput extends React.Component<Props, InputState<string>> {
-  schools: Array<string>;
-
-  constructor(props: Props) {
-    super(props);
-    this.schools = props.schools;
-    this.state = { value: '' };
-  }
-
-  doesStringMatch(item, value) {
-    if (value !== '') {
-      if (isUpper(value)) {
-        const firstLetters = item.split(" ").map(word => word[0]);
-
-        for (let char of [...value]) {
-          if (!firstLetters.includes(char)) {
-            return false;
-          }
-        }
-
-        return true;
-      }
-      else {
-        return item.toLowerCase().indexOf(value.toLowerCase()) !== -1;
-      }
-    }
-    else {
-      return false;
-    }
-  }
-
-  render() {
-    return (
-      <Autocomplete
-        getItemValue={(item) => item}
-        items={ this.schools }
-        shouldItemRender={ this.doesStringMatch }
-        renderItem={(item, isHighlighted) =>
-          <div style={{ background: isHighlighted ? 'lightgray' : 'white' }}>
-            { item }
-          </div>
-        }
-        value={ this.state.value }
-        onChange={(e) => this.setState({ value: e.target.value })}
-        onSelect={(value) => this.setState({ value })}
-      />
-    );    
-  }
-};
 
 class ApplyPage extends React.Component<Props, ApplyPageState> {
   unmounted: boolean;
